Added auth

+++ frontend/src/components/workoutDetails.js
@@ -1,31 +1,40 @@
-import {useWorkoutsContext} from '../hooks/useWorkoutsContext'
+import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
+
+// date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
-    const {dispatch} = useWorkoutsContext()
+  const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
 
-    const handleClick = async () => {
-        const response = await fetch(`/api/workouts/${workout._id}`, {
-            method: 'DELETE'
-        })
-        const json = await response.json()
-        
-        if (response.ok) {
-            dispatch({type: 'DELETE_WORKOUT', payload: json})
-        } else {
-            console.log(json.error)
-        }
+  const handleClick = async () => {
+    if (!user) {
+      return
     }
 
-    return (
-      <div className="workout-details">
-        <h4>{workout.title}</h4>
-        <p><strong>Load (kg): </strong>{workout.load}</p>
-        <p><strong>Number of reps: </strong>{workout.reps}</p>
-        <p>{formatDistanceToNow(new Date(workout.createdAt),{addSuffix: true})}</p>
-        <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
-      </div>
-    )
+    const response = await fetch('/api/workouts/' + workout._id, {
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
+    })
+    const json = await response.json()
+
+    if (response.ok) {
+      dispatch({type: 'DELETE_WORKOUT', payload: json})
+    }
   }
-  
-  export default WorkoutDetails
+
+  return (
+    <div className="workout-details">
+      <h4>{workout.title}</h4>
+      <p><strong>Load (kg): </strong>{workout.load}</p>
+      <p><strong>Reps: </strong>{workout.reps}</p>
+      <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
+      <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
+    </div>
+  )
+}
+
+export default WorkoutDetails
